updater: handle missing inactive partition set
If a variant is built without support for in-place updates, it will
not have the second, "inactive" set of partitions.

Since `signpost` is required for various operations, like marking a
boot successful, or indicating that the initial boot has finished, it
cannot be omitted entirely. Teach it to deal with only one partition
set, by making certain operations either optional or fallible.

Signed-off-by: Ben Cressey <bcressey@amazon.com>