kubernetes: move dockershim link to relative
The link created for dockershim.sock was an absolute path which worked
fine until that link is mounted into a container under a different path.
This would break use cases where the container needs to follow that link
to get to the containerd socket. This commit moves to a relative symlink
so that this will work via paths such as /host/run/dockershim.sock.