datastore: limit to_pairs input to serde_json::Value
This will simplify the input space for the serializer, allowing us to be
confident that we can make changes to the serialize_{type} methods
knowing that the space is testable.

This also sets us up well for settings extensions, as in that world we
won't have access to a Settings struct, so we will use serde_json::Value
as the de facto type for settings in Bottlerocket.

Signed-off-by: Sam Berning <bernings@amazon.com>