datastore: skip serializing null values instead of erroring
Now that we've limited the scope of the Serializer input to just
serde_json::Value, we can be confident that serialize_unit() would only
be called on a null value.

We can now define the behavior for that, instead of throwing a "bad
type" error. In this case, we prefer to omit the key/value pair from the
serialization output because we'd rather not persist null values in our
datastore, and the behavior should be the same on deserialization.

This should make creating settings models for Bottlerocket less
error-prone, as it's no longer necessary to add serde's
`skip_serializing_if` macro to all of the Option<T>s in the model. This
makes that behavior default for Option<T>.

Signed-off-by: Sam Berning <bernings@amazon.com>