kubernetes: handle null in credential provider template
Prior to [bottlerocket-settings-sdk@3c24f0e], the credential providers
would have explicitly returned null from APIServer for the Option<T>s in
the CredentialProvider struct. The `or` helper used in this template
behaves correctly if `this.environment` is explicitly null, but not if
it's unset.

This should correctly handle an unset `this.environment`.

[bottlerocket-settings-sdk@3c24f0e]: https://github.com/bottlerocket-os/bottlerocket-settings-sdk/commit/3c24f0eaf7229937adcd1ac560bd83e80474a1e9

Signed-off-by: Sam Berning <bernings@amazon.com>