driverdog: fix module load bug
The expected behavior for link-modules and load-modules is to load all
the modules from all the driver configuration files unless a specific
modules set is specified. In the case that a modules set is specified,
then driverdog should only link or load modules from the specified
target configuration. The link logic worked this way but the load logic
would always run the else clause, therefore always loading all modules
even if only one was specified.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>