apiserver: use direct paths for ephemeral storage
The path for ephemeral storage devices is created by udev on early boot
after ephemeral-storage init is called but if you call ephemeral-storage
bind immediately after this, the path may not exist, causing the
resulting commands to fail. This commit switches to a path that is known
to exist to avoid this issue.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>