add: nvidia-cdi-hook to nvidia-container-toolkit
For full CDI support, the `nvidia-container-toolkit` package should
include the `nvidia-cdi-hook`. The hook nvidia-cdi-hook is the CLI
tool that is expected to be called by the container runtime, when
specified by the CDI file.

For more information about how this hook is packaged by NVIDIA, see
the official RPM [1]. For how the CDI hook interacts with the
container runtime, see the CDI hook documentation [2].

[1]: https://github.com/NVIDIA/nvidia-container-toolkit/blob/main/packaging/rpm/SPECS/nvidia-container-toolkit.spec
[2]: https://github.com/NVIDIA/nvidia-container-toolkit/blob/main/cmd/nvidia-cdi-hook/README.md