ghostdog: add match-nvidia-driver argument
ghostdog can now look up PCI devices and confirm which NVIDIA driver
should be used. The match-nvidia-driver takes one argument, which type
of driver such as tesla or open-gpu and exits 0 if that matches the PCI
devices currently present. If those devices do not match the provided
driver, it exits 1. This can be used to set ExecCondition for things
like linking or loading drivers.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>