kmod-5.15-nvidia: move kernel module services into package
The os package doesn't need to concern itself with NVIDIA specific
loading behavior. It will provide driverdog, but the configurations read
by driverdog will be included with the specific kernel modules package
that provides the drivers described in the configuration. This moves the
tesla and open-gpu services into the kmod-5.15-nvidia package instead.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>