ecs-gpu-init: fix unit dependencies on GPU drivers
The GPU driver functionality has split into two units, one for the
proprietary driver and one for the open one. Depending on the version
of the driver, one or both units might be present.

Update the dependencies to reflect this. Also drop the "Requires", as
the "After" ordering should be sufficient, and "Requires" will cause
a failure if the unit is not installed.

Signed-off-by: Ben Cressey <bcressey@amazon.com>