kernel-6.1: exclude more object files from devel
Now that an unpacked set of kernel-devel files is present in the
buildroot, object files will be found by the `/usr/lib/rpm/brp-strip`
invocation that tries to ensure all files are stripped.

This can lead to build failures if the host and target arch don't
match, and if the target arch requires a host tool to be built. The
object files will be built for the host, and the target's `strip`
command will not recognize the format.

In any case, these object files shouldn't be included, as they may
need to be rebuilt to match the running host architecture at the time
the kernel-devel files are used.

Fixes: 76af8bef ("kernel-6.1: also provide uncompressed devel files")

Signed-off-by: Ben Cressey <bcressey@amazon.com>