login: start the getty services earlier
The login package is intended to help debug failures that occur early
in boot, but by default the getty services are not started until
`multi-user.target`, which is only activated if everything works as
expected.

Start the `getty.target` as part of `preconfigured.target` so that
the login prompt will be available even if there are problems with
networking or filesystem mounts.

Signed-off-by: Ben Cressey <bcressey@amazon.com>