selinux-policy: adjust kernel permissions for NFS
When serving files over NFS, the kernel's permissions to access the
inode are checked when determining whether the file can be executed
by clients. If the kernel lacks permission, then the client will see
a permission error and won't be able to execute the program.

Work around this by allowing the kernel access to execute mutable
files, while denying it permission to execute a new program without
transitioning to a new SELinux context.

To close off other potential paths to execution, define a transition
to a "forbidden" type that is explicitly denied, and deny the use of
any object as an entry point to that domain.

Signed-off-by: Ben Cressey <bcressey@amazon.com>