kernel: disable BPF preload and bpfilter helpers
The kernel build will automatically enable these helpers if it finds
that the target C compiler can build a userspace program that depends
on the C library. However, the helpers require additional libraries
such as libelf which are not guaranteed to be present.

Since we don't make use of precompiled BPF programs, disable these
config options rather than leaving it up to auto-detect.

Signed-off-by: Ben Cressey <bcressey@amazon.com>