Bloodhound: Add warning for 3.4.1.1 check fail in K8s
This test is expected to fail on Kubernetes variants as
Kubernetes needs the iptables rule -P FORWARD ACCEPT for its
operation and it is not recommended to modify this rule as it
could lead to adverse effects of service operation. This
rule exists in Bottlerocket because it is possible to run
Bottlerocket with default deny (on ECS for instance).