driverdog: Add ability to detect and decompress objects
The files that are linked at build time could be compressed which allows
space savings on disk for a minor boot time increase. driverdog now can
detect that the requested file may be compressed and have a .gz
extension. If so, then it will decompress before working with the file.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>