sources: add rottweiler
rottweiler provides a uniform wrapper around the standard underlying
tools: systemd-creds, cryptsetup, and the kernel's fscrypt API.

It requires a TPM2 device, and uses systemd-creds to generate wrapped
keys that are sealed to specific PCRs.

It doesn't support any form of interactive authentication or fallback
to keys stored outside the TPM2.

It also provides "measure" commands to ensure consistent measurements
of the kernel command line, OS settings, and boot phases.

Signed-off-by: Ben Cressey <bcressey@amazon.com>