release: support encryption for BOTTLEROCKET-DATA
Add an "encrypt" unit that checks whether BOTTLEROCKET-DATA has been
encrypted; if not, encrypt it via `rottweiler` using a key sealed to
the TPM2 device.

Add an "unlock" unit based on what `systemd-cryptsetup-generator`
would create. The generator relies on parsing `/etc/crypttab`, which
happens very early in boot. Writing `/etc/crypttab` would require a
systemd daemon reload afterwards to re-run the generator, which adds
latency and would require a separate stage of boot to accomplish the
proper sequencing.

Override subsequent services via drop-in so that the correct device
is mounted, resized, and grown. This must be the "attached", unlocked
LUKS device rather than the underlying block device.

Signed-off-by: Ben Cressey <bcressey@amazon.com>