release: support encryption for BOTTLEROCKET-PRIVATE
Add an "encrypt" unit that checks whether the datastore directory on
BOTTLEROCKET-PRIVATE has been encrypted; if not, encrypt it via
`rottweiler` using a key sealed to the TPM2 device.

Add an "unlock" unit that loads the encryption key into the kernel,
which allows files to be written and read from the directory.

Signed-off-by: Ben Cressey <bcressey@amazon.com>