release: measure kernel command line and settings
The kernel command line as viewed from userspace is measured into PCR
9, which is the standard destination.

Normally the kernel performs the measurement, but since Bottlerocket
allows the command line to be customized through bootconfig, the
kernel measurement is disabled so that the final value that's parsed
by the kernel can be measured instead.

After the initial round of settings have been applied, the effective
OS settings are measured into PCR 8. This happens before bootstrap
commands or bootstrap containers run, either of which could modify
settings. However, the measured settings will reflect the intent to
use these mechanisms, which could be obscured if the measurement took
place at a later point.

PCR 8 is normally used by GRUB, but our GRUB build does not enable
the TPM2 module, and never has, so the PCR is available for our own
measurements.

Signed-off-by: Ben Cressey <bcressey@amazon.com>